refactor: use core logger in examples (#97)
