+++ examples/nip46/unified-example.ts
@@ -21,7 +21,7 @@ import {
   generateKeypair,
   verifySignature,
 } from "../../src";
-import { LogLevel } from "../../src/nip46/utils/logger";
+import { LogLevel } from "../../src/utils/logger";
 import { NostrRelay } from "../../src/utils/ephemeral-relay";
 
 async function main() {
